wu ming

+++ app/src/main/java/com/cs165/domefavor/domefavor/TaskItem.java
@@ -78,7 +78,7 @@ public class TaskItem {
         this.content = content;
     }
 
-    void setPrice(int price){
+    void setPrice(double price){
         this.price = price;
     }
